% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{singlechoice_opts}
\alias{singlechoice_opts}
\alias{multichoice_opts}
\title{Get options for single and multi choice questions}
\usage{
singlechoice_opts(metadata)

multichoice_opts(metadata)
}
\arguments{
\item{metadata}{data.frame containing the metadata}
}
\value{
data.frame with variables \code{var} (variable), \code{label} (the
variable label), \code{vals} (possible values for the variable) and
\code{labs} (the labels related to each value in \code{vals})

data.frame with variables \code{ovar} (the variable as it appears in
the data dictionary/metadata), \code{var} (the variable as it appears in
the data itself), \code{vlabel} (the variable label), \code{vals} (possible
values for the variable) and \code{labs} (the labels related to each value in
\code{vals})
}
\description{
Get options for single and multi choice questions
}
\details{
Multiple choice variables exist in REDCap data as a set of 0/1/TRUE/FALSE
variables, where 1/TRUE represents a selected/checked answer. Hence, for a
single multiple choice 'question' in the datadictionary/metadata with \code{n} options,
there are \code{n} variables. Each variable is the variable name (e.g. morbidities)
followed by 3 underscores (\code{___}) and the option number (e.g. 1) -
\code{morbidities___1}.
}
