% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_logical.R
\name{convert_to_logical}
\alias{convert_to_logical}
\title{Convert variables to logical}
\usage{
convert_to_logical(
  data,
  vars,
  true = "Yes",
  replace = TRUE,
  append = "_logical"
)
}
\arguments{
\item{data}{dataframe}

\item{vars}{character string of variables to convert}

\item{true}{value which should become \code{TRUE}}

\item{replace}{Replace the indicated variables}

\item{append}{text to append to new variables (when \code{replace = TRUE})}
}
\value{
\code{data} with modified variables, potentially with additional variables
(if \code{replace = TRUE})
}
\description{
This is particularly useful for binary variables that have been encoded with
e.g. Yes and No as options. Variable labels are retained, which may or may not
make sense, depending on the variable
}
\examples{
data(mtcars)
convert_to_logical(mtcars, "am", 1)
convert_to_logical(mtcars, c("am", "vs"), 1)
convert_to_logical(mtcars, c("am", "vs"), 1, FALSE)
convert_to_logical(mtcars, c("am", "vs"), 1, FALSE, "_lgl")
}
