% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_checkboxes.R
\name{count_checkboxes}
\alias{count_checkboxes}
\title{Count checkbox responses}
\usage{
count_checkboxes(data, ..., check_text = "Checked", negate = FALSE,
  key_text = "field_name", value_text = "n", group = NULL)
}
\arguments{
\item{data}{A tbl_df or data.frame}

\item{...}{Checkbox column names. Supports `dplyr` `select_helpers`}

\item{check_text}{Character string of the response to count. Default is
"Checked"}

\item{negate}{Boolean. If `TRUE`, the count is `!= check_text`. Default is
`FALSE`}

\item{key_text}{Character string to name result column of the columns}

\item{value_text}{Character string to name the result column of counts}

\item{group}{Column name to `group_by`}
}
\value{
a tble_df
}
\description{
Get summaary counts easily from checkbox style variables/columns in a REDCap
data set.
}
\examples{
library(tibble)
df <- tibble::tibble(
  id = c(1:100),
  group = sample(c("A", "B", "C"), size = 100, replace = TRUE),
  preferred_color___1 = sample(c("Checked", "Unchecked"), size = 100, replace = TRUE),
  preferred_color___2 = sample(c("Checked", "Unchecked"), size = 100, replace = TRUE),
  preferred_color___3 = sample(c("Checked", "Unchecked"), size = 100, replace = TRUE),
  preferred_color___4 = sample(c("Checked", "Unchecked"), size = 100, replace = TRUE),
  preferred_color___5 = sample(c("Checked", "Unchecked"), size = 100, replace = TRUE))

count_checkboxes(data = df,
                 preferred_color___1:preferred_color___5)

count_checkboxes(data = df,
                 preferred_color___1, preferred_color___5)

count_checkboxes(data = df,
                 dplyr::starts_with("preferred_color"))

count_checkboxes(data = df,
                 preferred_color___1:preferred_color___5,
                 check_text = "Unchecked",
                 key_text = "checkbox",
                 value_text = "count")

count_checkboxes(data = df,
                 preferred_color___1:preferred_color___5,
                 group = group)

count_checkboxes(data = df,
                 preferred_color___1:preferred_color___5,
                 negate = TRUE,
                 group = group)
}
