% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{rc_prep}
\alias{rc_dates}
\alias{rc_datetimes}
\alias{split_by_form}
\title{Deprecated functions
These functions have been renamed to be more consistent with the rest of the package. They may be removed in a future version.}
\usage{
rc_prep(
  data,
  metadata,
  rep = FALSE,
  rep_date = rep,
  rep_datetime = rep,
  rep_singlechoice = rep,
  rep_multichoice = rep,
  app_date = "_date",
  app_datetime = "_datetime",
  app_singlechoice = "_factor",
  app_multichoice = "_factor",
  ...
)

rc_dates(data, metadata, replace = FALSE, append = "_date")

rc_datetimes(data, metadata, replace = FALSE, append = "_datetime", ...)

split_by_form(data, metadata)
}
\arguments{
\item{data}{dataframe}

\item{metadata}{datadictionary as exported from REDCap or downloaded from the API}

\item{rep}{replace variables. If FALSE, encoded versions of the variable will be created}

\item{rep_date, rep_datetime, rep_singlechoice, rep_multichoice}{replace the indicated variable type}

\item{app_date, app_datetime, app_singlechoice, app_multichoice}{text to append to the newly generated variables name (if \code{rep_*} is FALSE)}

\item{...}{options passed to/from other methods}

\item{replace}{indicator of whether to replace original variables or not}

\item{append}{text to append to the newly generated variables name (if \code{replace} is TRUE)}
}
\value{
list of dataframes
}
\description{
Deprecated functions
These functions have been renamed to be more consistent with the rest of the package. They may be removed in a future version.
}
\section{Functions}{
\itemize{
\item \code{rc_prep()}: original function name for \code{redcap_prep}

\item \code{rc_dates()}: original function name for \code{redcap_dates}

\item \code{rc_datetimes()}: original function name for \code{redcap_datetimes}

\item \code{split_by_form()}: deprecated in favour of \code{redcap_toform}
Split a manually exported REDCap dataset into forms

}}
