% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vartypes.R
\name{get_vartypes}
\alias{get_vartypes}
\title{Get Variable Type for downloading data}
\usage{
get_vartypes(meta)
}
\arguments{
\item{meta}{project meta information as downloaded by redcap_export_meta}
}
\value{
Character vector of variable classes.
}
\description{
This function uses the project meta information to determine a vector of
variable classes. This vector can be used to passed into read_csv when
downloading data via API to make sure that the classes are read correctly.
Checkbox, radio, dropdown and yesno items are set as integer, text fields
are set to integer, numeric or character depending on their field validation,
calculated fields are set to numeric. Every other variable type are set as
character.
}
\examples{
# data(proj)
# get_vartypes(proj)

# if using local data:
# token <- "xxxxx"
# url <- "xxxxx"
# meta <- redcap_export_meta(token,url)
# get_vartypes(meta)
}
