% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importdemo}
\alias{importdemo}
\title{Example import data}
\format{
A data frame with 17 observations and 32 variables.
}
\usage{
importdemo
}
\description{
Simulated import data for the most common REDCap field types and validations
to be used with the data dictionary \code{meta} for testing of
\code{redcap_select_rename} and \code{redcap_recode}. Item names and coding
show mismatches on purpose so that the interactive functions can be tested.
}
\keyword{datasets}
