% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importdemo_dict}
\alias{importdemo_dict}
\title{Example data dictionary}
\format{
A data frame with 38 variable specifications.
}
\usage{
importdemo_dict
}
\description{
Simulated data dictionary for the most common REDCap field types and
validations to be used with the sample data \code{importdemo_data} for testing of
\code{redcap_import_select} and \code{redcap_import_recode}. Item names and coding
show mismatches on purpose so that the interactive functions can be tested.
}
\keyword{datasets}
