% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{label_others}
\alias{label_others}
\title{Label non-single/multiple choice/date(time) fields
\code{singlechoice_factor}, \code{multichoice_factor}, \code{rc_date} and \code{rc_datetime}}
\usage{
label_others(data, metadata)
}
\arguments{
\item{data}{dataframe}

\item{metadata}{redcap data dictionary}
}
\description{
Label non-single/multiple choice/date(time) fields
\code{singlechoice_factor}, \code{multichoice_factor}, \code{rc_date} and \code{rc_datetime}
}
