% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Example data dictionary}
\format{
A data frame with 32 variable specifications.
}
\usage{
meta
}
\description{
Simulated data dictionary for the most common REDCap field types and
validations to be used with the sample data \code{importdemo} for testing of
\code{redcap_select_rename} and \code{redcap_recode}. Item names and coding
show mismatches on purpose so that the interactive functions can be tested.
}
\keyword{datasets}
