% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{multichoice_factor}
\alias{multichoice_factor}
\title{create factors for multiple choice variables}
\usage{
multichoice_factor(data, metadata, replace = FALSE, append = "_factor")
}
\arguments{
\item{data}{the data.frame to modify}

\item{metadata}{metadata/datadictionary}

\item{replace}{whether to overwrite the existing data .}

\item{append}{text to append to the variable name if not overwriting}
}
\value{
input data.frame with additional factor variables.
}
\description{
Converts the numeric values returned from REDCap to factors (with levels Yes/No).
This function also applies labels to the variable itself, based on the option label.
}
