% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proj}
\alias{proj}
\title{Example meta data}
\format{
A list of 3 with metadata (data dictionary), instrument (CRF specifications),
and project (global project settings).
}
\usage{
proj
}
\description{
Simulated complete meta data for the most common REDCap field types and
validations.
}
\keyword{datasets}
