% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{rc_dates}
\alias{rc_dates}
\title{Convert dates stored as strings to \code{Date} variables}
\usage{
rc_dates(data, metadata, replace = FALSE, append = "_date")
}
\arguments{
\item{data}{the data.frame to modify}

\item{metadata}{metadata/datadictionary}

\item{replace}{whether to overwrite the existing data .}

\item{append}{text to append to the variable name if not overwriting}
}
\value{
input data.frame with additional date variables/variables converted to dates.
}
\description{
Converts the numeric values returned from REDCap to factors (with levels Yes/No).
This function also applies labels to the variable itself, based on the option label.
}
