% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{rc_prep}
\alias{rc_prep}
\title{Convert REDCap variable types (dates, datetimes, factors)}
\usage{
rc_prep(data, metadata)
}
\arguments{
\item{data}{dataframe}

\item{metadata}{data dictionary from REDCap}
}
\value{
dataframe with converted factors, dates, POSIX, ...
}
\description{
Convert REDCap variable types (dates, datetimes, factors)
}
