% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_dates.R
\name{redcap_dates}
\alias{redcap_dates}
\title{REDCap Date Conversion}
\usage{
redcap_dates(var)
}
\arguments{
\item{var}{variable to convert}
}
\value{
converted variable
}
\description{
This function is trying to take all (most)
possibilities to enter dates into account and converts them to as.Date
format.
}
\examples{
var <- "dataframe$var"
redcap_dates(var)
}
