% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_export.R
\name{redcap_export_batch}
\alias{redcap_export_batch}
\title{Export data in batches}
\usage{
redcap_export_batch(
  token,
  url,
  batchsize = 1000,
  meta = NULL,
  byform = FALSE,
  remove_empty = TRUE,
  ...
)
}
\arguments{
\item{token}{REDcap API token}

\item{url}{address of the API}

\item{batchsize}{number of records per batch}

\item{meta}{metadata from \code{redcap_export_meta} (will be downloaded if not provided)}

\item{byform}{logical. Download data by form (see \link{redcap_export_byform})}

\item{remove_empty}{when using byform: should empty rows be removed from the dataset (REDCap automatically
creates all forms for an event when any form in the event is created)}

\item{...}{other parameters passed to the API (see your REDCap API documentation for options)}
}
\value{
depending on \code{byform}, either a list of dataframes or a single dataframe
}
\description{
Exports of large databases may fail using the standard export methods implemented
in \link{redcap_export_tbl} and \link{redcap_export_byform}. To remedy this,
the \code{redcap_export_batch} function exports data in smaller chunks (of 1000
records by default)
}
\examples{
# token <- "some_really_long_string_provided_by_REDCap"
# as a single dataframe
# redcap_export_batch(token, "https://www.some_redcap_url.com/api/")
# as a list of dataframes (forms)
# redcap_export_batch(token, "https://www.some_redcap_url.com/api/", byform = TRUE)
}
\seealso{
\link{redcap_export_tbl}, \link{redcap_export_byform}
}
