% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_export.R
\name{redcap_export_tbl}
\alias{redcap_export_tbl}
\title{Export tables from REDCap}
\usage{
redcap_export_tbl(token, url, content, ...)
}
\arguments{
\item{token}{REDcap API token}

\item{url}{address of the API}

\item{content}{content to download}

\item{...}{other parameters passed to the API (see your REDCap API helpfile for options)}
}
\value{
dataframe
}
\description{
Export tables from REDCap
}
\examples{
# token <- "some_really_long_string_provided_by_REDCap"
# redcap_export_tbl(token, "https://www.some_redcap_url.com/api/", "record")
}
