% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_times.R
\name{redcap_import_times}
\alias{redcap_import_times}
\title{REDCap Time Conversion}
\usage{
redcap_import_times(var, unk_min = "00", unk_sec = "00")
}
\arguments{
\item{var}{variable to convert}

\item{unk_min}{Minutes to use if unknown. The default is 00 (11 -> 11:00).}

\item{unk_sec}{Seconds to use if unknown. The default is 00 (11:11 -> 11:11:00).}
}
\value{
converted variable
}
\description{
This function is used to prepare time-values in a data table for import in REDCap.
In an Excel, time-values can be entered in various shapes and forms. This function
tries to take most possibilities how time-values could have been entered into
account and converts them to a format that can be imported into REDCap.
}
\examples{
var <-c("11:11:00","11:11","11")
redcap_import_times(var)
}
