% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_select_rename.R
\name{redcap_select_rename}
\alias{redcap_select_rename}
\title{REDCap Select and Rename}
\usage{
redcap_select_rename(
  import_data,
  rc_token,
  url = "https://redcap.ctu.unibe.ch/api/",
  dict = NULL,
  forms = NULL
)
}
\arguments{
\item{import_data}{data to be imported}

\item{rc_token}{REDCap API token}

\item{url}{link to REDCap API}

\item{dict}{data dictionary (e.g. as downloaded from REDCap or via
\code{redcap_export_meta(rc_token, url)$meta}). If not supplied, this will
be downloaded from the API using \code{rc_token}.}

\item{forms}{list of REDCap forms of which variable names will be displayed. Default = all forms.}
}
\value{
???
}
\description{
REDCap Select and Rename
}
\examples{
token <- "xxxxx"
file <- "inst/data.csv"
redcap_select(file, token)
}
