% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcaptools.R
\docType{package}
\name{redcaptools}
\alias{redcaptools}
\title{redcaptools: Misc tools for working with REDCap data in R}
\description{
This package contains a some functions that I find handy when working with
REDCap data.
}
\examples{
# Example usage:
library(redcaptools)

}
