% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{singlechoice_factor}
\alias{singlechoice_factor}
\title{create factors for single choice variables}
\usage{
singlechoice_factor(data, metadata, replace = FALSE, append = "_factor")
}
\arguments{
\item{data}{the data.frame to modify}

\item{metadata}{metadata/datadictionary}

\item{replace}{whether to overwrite the existing data .}

\item{append}{text to append to the variable name if not overwriting}
}
\value{

}
\description{
Converts the numeric values returned from REDCap to factors. This function also
applies labels to the variable itself.
}
