% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_form.R
\name{split_by_form}
\alias{split_by_form}
\title{Split a manually exported REDCap dataset into forms}
\usage{
split_by_form(data, metadata)
}
\arguments{
\item{data}{dataframe}

\item{metadata}{datadictionary as exported from REDCap or downloaded from the API}
}
\value{
list of dataframes
}
\description{
Split a manually exported REDCap dataset into forms
}
