% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{rc_prep}
\alias{rc_prep}
\title{Convert REDCap variable types (dates, datetimes, factors) and apply labels}
\usage{
rc_prep(
  data,
  metadata,
  rep = FALSE,
  rep_date = rep,
  rep_datetime = rep,
  rep_singlechoice = rep,
  rep_multichoice = rep,
  app_date = "_date",
  app_datetime = "_datetime",
  app_singlechoice = "_factor",
  app_multichoice = "_factor",
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{metadata}{data dictionary from REDCap}

\item{rep}{replace variables. If FALSE, encoded versions of the variable will be created}

\item{rep_date, rep_datetime, rep_singlechoice, rep_multichoice}{replace the indicated variable type}

\item{app_date, app_datetime, app_singlechoice, app_multichoice}{text to append to the newly generated variables name (if \code{rep_*} is FALSE)}

\item{...}{options passed to/from other methods}
}
\value{
dataframe with converted factors, dates, POSIX, ...
}
\description{
Convert REDCap variable types (dates, datetimes, factors) and apply labels
}
