% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_dates.R
\name{redcap_dates}
\alias{redcap_dates}
\title{REDCap Date Conversion}
\usage{
redcap_dates(var, unk_day = "01", unk_month = "01", unk_cent = "20")
}
\arguments{
\item{var}{variable to convert}

\item{unk_day}{Day to use if unknown, i.e. if only the year or only the month
+ year is found. The default is 01 (2022 -> 2022-01-01).}

\item{unk_month}{Month to use if unknown, i.e. if only the year is found. The
default is 01 (2022 -> 2022-01-01).}

\item{unk_cent}{Century to use if unknown, i.e. if only the year is found.
The default is 20 (22 -> 2022)}
}
\value{
converted variable
}
\description{
This function is used to prepare date fields in a data table for import in
REDCap. The function tries to take all (most)possibilities to enter dates
into account and converts them to as.Date format.
}
\examples{
var <-c("01.12.2022", "12.2022", "2022", "01/12/2022")
redcap_dates(var)
}
