% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_export.R
\name{redcap_export_meta}
\alias{redcap_export_meta}
\title{Export the most important REDCap metadata tables}
\usage{
redcap_export_meta(
  token,
  url,
  tabs = c("metadata", "event", "formEventMapping", "instrument"),
  ...
)
}
\arguments{
\item{token}{REDcap API token}

\item{url}{address of the API}

\item{tabs}{tables to export. \code{project} is always added.}

\item{...}{other parameters passed to the API (see your REDCap API documentation for options)}
}
\value{
list of dataframes
}
\description{
The REDCap API has a large number of API endpoints. Those that are metadata-type
details are listed on this page. The
}
\details{
Allowed tabs are
\itemize{
\item \code{arm} - labels of a projects arms
\item \code{dag} - data access groups (DAGs)
\item \code{userDagMapping} - mapping between users and DAGs
\item \code{event} - list of events in the project (only
available for longitudinal projects)
\item \code{exportFieldNames} - list of the fields that the API returns
\item \code{instrument} - list of instruments (eCRFs/forms) in the project
\item \code{formEventMapping} - mapping between instruments (forms) and events (only
available for longitudinal projects)
\item \code{metadata} - the data dictionary
\item \code{project} - information on the project
\item \code{record} - the data itself. The method has many options. See the API help
page on your REDCap instance
\item \code{repeatingFormsEvents} - which forms can repeat on which events
\item \code{report} - access custom reports defined in REDCap
\item \code{version} - REDCap version
\item \code{user} - list of users
\item \code{userRole} - rights for each role
\item \code{userRoleMapping} - user-roll mapping
}
}
\note{
tables that are not relevant for non-longitudinal projects (e.g.
formEventMapping and event) are silently removed
}
\examples{
# token <- "some_really_long_string_provided_by_REDCap"
# redcap_export_meta(token, "https://www.some_redcap_url.com/api/")
}
