% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import_dates_times.R
\name{redcap_import_datetime}
\alias{redcap_import_datetime}
\title{REDCap Date-Time Conversion}
\usage{
redcap_import_datetime(
  var,
  args_rc_dates = list(),
  args_rc_times = list(),
  date_only = FALSE
)
}
\arguments{
\item{var}{Variable to convert}

\item{args_rc_dates}{List with arguments for \code{redcap_import_dates} (e.g.
args_rc_dates = list(unk_day = 01,format = "american"))}

\item{args_rc_times}{List with arguments for \code{redcap_import_times} (e.g.
args_rc_times = list(unk_min = 01,unk_sec = 01))}

\item{date_only}{If TRUE, only the date will be included in the output and
the time value will be removed. Default = FALSE.}
}
\value{
converted variable
}
\description{
This function prepares date-time values in a data table for import in REDCap.
It parses date and time values and processes them using
\code{redcap_import_dates} and \code{redcap_import_times}. This ensures that
date-time entries, which may have been entered in various formats in Excel,
are converted into a format compatible with REDCap.
}
\examples{
var <-c("1.2.24 11:11:00","1.2.22 11:11","1.2.24 11", "31341.34375")
redcap_import_datetime(var)

}
