% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import_dates_times.R
\name{redcap_import_times}
\alias{redcap_import_times}
\title{REDCap Time Conversion}
\usage{
redcap_import_times(var, unk_min = "00", unk_sec = "00")
}
\arguments{
\item{var}{Variable to convert}

\item{unk_min}{Minutes to use if unknown. The default is 00 (11 -> 11:00).}

\item{unk_sec}{Seconds to use if unknown. The default is 00 (11:11 ->
11:11:00).}
}
\value{
converted variable
}
\description{
This function prepares time values in a data table for import in REDCap. In
Excel, time values can be entered in various formats. This function attempts
for the most common ways time values may have been entered and converts them
into a format compatible with REDCap.
}
\examples{
var <-c("11:11:00","11:11","11","0.34375" )
redcap_import_times(var)
}
