% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_select_rename.R
\name{redcap_select_rename}
\alias{redcap_select_rename}
\title{REDCap Select and Rename}
\usage{
redcap_select_rename(
  import_data,
  dict = NULL,
  rc_token,
  rc_url,
  forms = NULL,
  auto = TRUE
)
}
\arguments{
\item{import_data}{Data frame to be imported}

\item{dict}{Data dictionary (e.g. as
downloaded from REDCap or via \code{redcap_export_meta(rc_token,
rc_url)$meta}). If not supplied, this will be downloaded from the API using
\code{rc_token} and \code{rc_url}.}

\item{rc_token}{REDCap API token}

\item{rc_url}{Link to REDCap API.}

\item{forms}{List of REDCap forms of which variable names will be displayed.
Default = all forms.}

\item{auto}{if TRUE, variables with matching names will be automatically
selected. If FALSE, the user can decide if the variable shall be imported or
not. Default = TRUE.}
}
\value{
Data frame with selected/renamed data. Log-file with executed code.
 CSV-table with overview.
}
\description{
This function loops through all the variable names of a data set and lets the
user compare them with the variable names set up in REDCap. An API token is
needed to download the variable names from REDCap.Variables with matching
names in REDCap can simply be selected without renaming, variables without
matching names can be selected and renamed. The function returns a data frame
with the selected/renamed variables, writes an overview csv-table, and a short
summary together with the executed code to a log-file for copy-pasting and
adjusting/reusing.
}
\examples{
# data(importdemo)
# data(meta)
# redcap_select_rename(importdemo, meta)

# if using local data:
# token <- "xxxxx"
# url <- "xxxxx"
# file <- "data.csv"
# redcap_select_rename(file, rc_token = token, rc_url = url)
}
